requirements.txt
pymupdf>=1.23.0
pyserial>=3.5
